Swap from S3 source of truth to api endpoint
